fix(router): 补齐路由回调类型
