+++ packages/taro-h5/src/index.ts
@@ -1,6 +1,6 @@
-import Taro from './taro/index'
+import Taro from './api/taro'
 
 export * from './api/index'
+export * from './api/taro'
 
 export default Taro
-export * from './taro/index'
