--- packages/taro-h5/src/taro/index.ts
@@ -1,126 +1,0 @@
-import Taro from '@tarojs/api'
-
-import {
-  history,
-  createRouter,
-  getCurrentPages
-} from '@tarojs/router'
-import { permanentlyNotSupport } from '../api/utils'
-
-const {
-  Behavior,
-  getEnv,
-  ENV_TYPE,
-  Link,
-  interceptors,
-  getInitPxTransform,
-  Current,
-  options,
-  eventCenter,
-  Events,
-  preload,
-  useDidShow,
-  useDidHide,
-  usePullDownRefresh,
-  useReachBottom,
-  usePageScroll,
-  useResize,
-  useShareAppMessage,
-  useTabItemTap,
-  useTitleClick,
-  useOptionMenuClick,
-  usePullIntercept,
-  useShareTimeline,
-  useAddToFavorites,
-  useReady,
-  useRouter
-} = Taro as any
-
-const taro: typeof Taro = {
-  // @ts-ignore
-  Behavior,
-  getEnv,
-  ENV_TYPE,
-  Link,
-  interceptors,
-  Current,
-  options,
-  eventCenter,
-  Events,
-  preload,
-  history,
-  createRouter,
-  getCurrentPages: getCurrentPages as unknown as typeof Taro.getCurrentPages,
-  useDidShow,
-  useDidHide,
-  usePullDownRefresh,
-  useReachBottom,
-  usePageScroll,
-  useResize,
-  useShareAppMessage,
-  useTabItemTap,
-  useTitleClick,
-  useOptionMenuClick,
-  usePullIntercept,
-  useShareTimeline,
-  useAddToFavorites,
-  useReady,
-  useRouter
-}
-
-const initPxTransform = getInitPxTransform(taro)
-
-const requirePlugin = permanentlyNotSupport('requirePlugin')
-
-const pxTransform = function (size) {
-  // @ts-ignore
-  const { designWidth } = taro.config
-  return Math.ceil((((parseInt(size, 10) / 40) * 640) / designWidth) * 10000) / 10000 + 'rem'
-}
-const canIUseWebp = function () {
-  const canvas = document.createElement('canvas')
-  return canvas.toDataURL('image/webp').indexOf('data:image/webp') === 0
-}
-
-taro.requirePlugin = requirePlugin
-taro.pxTransform = pxTransform
-taro.initPxTransform = initPxTransform
-// @ts-ignore
-taro.canIUseWebp = canIUseWebp
-
-export default taro
-
-export {
-  Behavior,
-  getEnv,
-  ENV_TYPE,
-  Link,
-  interceptors,
-  initPxTransform,
-  Current,
-  options,
-  eventCenter,
-  Events,
-  preload,
-  requirePlugin,
-  pxTransform,
-  canIUseWebp,
-  history,
-  createRouter,
-  getCurrentPages,
-  useDidShow,
-  useDidHide,
-  usePullDownRefresh,
-  useReachBottom,
-  usePageScroll,
-  useResize,
-  useShareAppMessage,
-  useTabItemTap,
-  useTitleClick,
-  useOptionMenuClick,
-  usePullIntercept,
-  useShareTimeline,
-  useAddToFavorites,
-  useReady,
-  useRouter
-}
